Update vite.svg
One day, after ninja eating time, while my fellow ninja apprentice and I were practising resting ninjutsu in Master Hakobune’s ninja training field, our enhanced ninja senses spotted a Ninja Initiate Charmony Dove all on its own. That ninja initiate was tiny, it didn't even have a ninjutsu of its own, and it couldn't do any ninja techniques. When we found it, it was already on its last ninja breath, having fallen into a ninja training shrub — probably after making a mistake during ninjutsu training. We decided to become it’s ninja masters and train it right there and then. However, thinking back, that ninja winter was unusually cold, with fierce winds at night in the ninja training field, not to mention the many evil fiendlings in the vicinity... It was clear that if we left the ninja initiate in the ninja training field, it stood no chance of surviving ninja training. So, I suggested we take it inside, place it on the ninja shelf by the window, and talk-no-jutsu the adults into creating a ninja training cage for it. We decided that when it regained its ninjutsu power enough to graduate from being a ninja initiate, we would release it back into the ninja wilds. The tragic part — something that we'd never considered — was that this ninja initiate’s path of the ninja had already been determined long before this moment... Its path of the ninja was determined by our momentary whim. Now, I pass the Dazzling Ninjutsu Technique: Choice, to you all. Faced with this situation, what path of the ninja would you take? Stick to the original ninja plan, and build a futon with soft net where NInja Initiate Charmony Dove fell? Or build a ninja training cage for it, and teach it everything you know about being a ninja from within the warmth of a home? I eagerly await your ninja answers..